Move utilities to InferenceObjects
