+++ test/test_utils.jl
@@ -43,22 +43,6 @@ pandas = ArviZ.pandas
         @test ArviZ.frompytype([[x2]]) == [[x]]
     end
 
-    @testset "rekey" begin
-        orig = (x=3, y=4, z=5)
-        keymap = (x=:y, y=:a)
-        @testset "NamedTuple" begin
-            new = @inferred NamedTuple ArviZ.rekey(orig, keymap)
-            @test new isa NamedTuple
-            @test new == (y=3, a=4, z=5)
-        end
-        @testset "Dict" begin
-            orig_dict = Dict(pairs(orig))
-            new = @inferred ArviZ.rekey(orig_dict, keymap)
-            @test new isa typeof(orig_dict)
-            @test new == Dict(:y => 3, :a => 4, :z => 5)
-        end
-    end
-
     @testset "topandas" begin
         @testset "DataFrames.DataFrame -> pandas.DataFrame" begin
             columns = [:a, :b, :c]
